Move coordinator e2e back to coordinator/internal/e2e/
The coordinator's own e2e package was incorrectly flattened into
coordinator/e2e/ alongside the repo-root e2e/ testbed suite.
Restore it to coordinator/internal/e2e/ where it belongs.
