+++ coordinator/api/consumer.go
@@ -26,7 +26,7 @@ import (
 	"time"
 
 	"github.com/eigeninference/d-inference/coordinator/auth"
-	"github.com/eigeninference/d-inference/coordinator/e2e"
+	"github.com/eigeninference/d-inference/coordinator/internal/e2e"
 	"github.com/eigeninference/d-inference/coordinator/payments"
 	"github.com/eigeninference/d-inference/coordinator/protocol"
 	"github.com/eigeninference/d-inference/coordinator/registry"
