+++ coordinator/api/sender_encryption_test.go
@@ -16,7 +16,7 @@ import (
 
 	"golang.org/x/crypto/nacl/box"
 
-	"github.com/eigeninference/d-inference/coordinator/e2e"
+	"github.com/eigeninference/d-inference/coordinator/internal/e2e"
 	"github.com/eigeninference/d-inference/coordinator/registry"
 	"github.com/eigeninference/d-inference/coordinator/store"
 )
