+++ coordinator/internal/e2e/coordinator_key.go
