+++ coordinator/internal/e2e/e2e/coordinator_key.go
